Add test case for add item file
