+++ __mocks__/loadList.js
@@ -1,0 +1,3 @@
+const loadList = () => Promise.resolve('new List');
+
+module.exports = loadList;
